pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Shyskov <shyskov@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.10.0"
djangorestframework = "^3.15.2"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
ruff = "^0.6.9"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
